Added discard support and zero write detection to system disk configuration
Adding the "discard=unmap" option to the system disk causes qemu to
unmap (i.e., convert into sparse extents) regions of the disk image
identified by the guest's use of the trim/discard command.

Writing zeroes to the free space on a virtual disk (to later shrink the
image) causes the disk image file to expand to its maximum size. Adding
"detect-zeroes=on" converts these to sparse writes for unallocated parts
of the disk.